Add mismatch leaves allowed structure tag
Any structure in this tag will be allowed to have leaves that do not match the logs. This allows players and other modders to tag structures where mismatched leaves/logs are intended and should not decay.

Name of tag can be changed. I added two bumblezone entries in the tag as an example. You can keep the entries or remove them. The tag entries are optional tag entries so they do not error even if bumblezone is not on.

Fixes https://github.com/TeamDiopside/SeparatedLeaves/issues/4